Refactor common and specific logic within the miner workers
